A branch that prints out everything written to the transceiver along with the time and the contents.  All writing is staggered so there should be no collisions.

+++ processors/renesas/src/common/flash.c
@@ -170,7 +170,7 @@ void clr_status_reg(UInt8 * addr);
 
     \return The number of bytes written
 */
-UInt16 write_data_flash(const UInt16 ADDR, const UInt8 * DATA, const UInt16 LEN)
+UInt16 write_data_flash(const UInt16 ADDR, const UInt8 * DATA, const UInt8 LEN)
 {
     UInt16 written = 0;
